+++ src/gemseo_mlearning/algos/opt/egobox/regression_spec.py
@@ -12,12 +12,23 @@
 # You should have received a copy of the GNU Lesser General Public License
 # along with this program; if not, write to the Free Software Foundation,
 # Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
-"""Settings of the optimization algorithms."""
+"""Trend strategy."""
 
 from __future__ import annotations
 
-from gemseo_mlearning.algos.opt.egobox.egor_settings import (  # noqa: F401
-    EGObox_Egor_Settings,
-)
-from gemseo_mlearning.algos.opt.sbo_settings import SBO_Settings  # noqa: F401
-from gemseo_mlearning.algos.opt.smt.ego_settings import SMT_EGO_Settings  # noqa: F401
+from enum import auto
+
+from strenum import StrEnum
+
+
+class RegressionSpec(StrEnum):
+    """The type of trend for the Gaussian process (GP) models."""
+
+    CONSTANT = auto()
+    """A constant trend."""
+
+    LINEAR = auto()
+    """A linear trend."""
+
+    QUADRATIC = auto()
+    """A quadratic trend."""